+++ src/include/smalloc.h
@@ -21,6 +21,6 @@ extern void *smrealloc(struct memnode *mp, void *ptr, size_t size);
 extern int sminit(struct memnode *mp, void *start, size_t size);
 extern int smdestroy(struct memnode *mp);
 extern size_t smget_free_space(struct memnode *mp);
-extern size_t smget_area_size(struct memnode *mp, void *ptr);
+extern int smget_area_size(struct memnode *mp, void *ptr);
 
 #endif
